New version 2.3.4. Added --cleanup=True/False and using it with rmtree()

+++ pilot.py
@@ -339,6 +339,11 @@ def get_args():
                             type=str2bool,
                             default=True,
                             help='Use HTTPS protocol for communications with server')
+    arg_parser.add_argument('--cleanup',
+                            dest='cleanup',
+                            type=str2bool,
+                            default=True,
+                            help='Cleanup work directory after pilot has finished')
 
     # Harvester and Nordugrid specific options
     arg_parser.add_argument('--input-dir',
@@ -439,7 +444,7 @@ def wrap_up(initdir, mainworkdir, args):
     exit_code = 0
 
     # cleanup pilot workdir if created
-    if initdir != mainworkdir:
+    if initdir != mainworkdir and args.cleanup:
         chdir(initdir)
         try:
             rmtree(mainworkdir)
